Fix a lot of stuff, idr

+++ kate.FileShare/Data/Models/UserLimitModel.cs
@@ -8,8 +8,8 @@ public class UserLimitModel
     public const string TableName = "UserLimits";
     public string UserId { get; set; }
     public UserModel User { get; set; }
-    public int? MaxFileSize { get; set; }
-    public int? MaxStorage { get; set; }
+    public long? MaxFileSize { get; set; }
+    public long? MaxStorage { get; set; }
 
     [DefaultValue(0)]
     public long SpaceUsed { get; set; }
